Add story for avatar component
- Minor UI changes to avatar component
