+++ package.json
@@ -42,9 +42,9 @@
         "@babel/runtime-corejs3": "^7.11.2",
         "@storybook/client-api": "^6.0.21",
         "@togglecorp/fujs": "^1.9.2-beta.0",
-        "file-saver": "^2.0.2",
         "d3-axis": "^2.0.0",
         "d3-scale": "^3.2.3",
+        "file-saver": "^2.0.2",
         "react-icons": "^3.11.0",
         "resize-observer": "^1.0.0"
     },
@@ -75,9 +75,9 @@
         "@storybook/addon-links": "^6.0.21",
         "@storybook/node-logger": "^6.0.21",
         "@storybook/react": "^6.0.21",
-        "@types/file-saver": "^2.0.1",
         "@types/d3-axis": "^2.0.0",
         "@types/d3-scale": "^3.1.0",
+        "@types/file-saver": "^2.0.1",
         "@types/node": "^14.11.2",
         "@types/react": "^16.9.49",
         "@types/react-dom": "^16.9.8",
