+++ src/components/Avatar/styles.css
@@ -20,12 +20,11 @@
         display: flex;
         align-items: center;
         justify-content: center;
-        border: var(--tui-width-separator-thin) solid var(--tui-color-separator);
         border-radius: 50%;
         width: 100%;
         height: 100%;
         text-align: center;
         text-transform: uppercase;
-        font-size: calc(0.6 * var(--size));
+        font-size: calc(var(--size) / 2);
     }
 }
